Restructure the renderer: Writer-based, one module per event type
Rendering now writes into a caller-supplied buffer in the style of
Display/Debug formatters: each event type implements the Event trait
(classify + render as one concept) and writes its headline and detail
lines through a Block writer, which owns the layout invariants
(detail-line indent and dimming, headline trimming, the truncated
source's overflow line). The per-piece format!-into-String
intermediates, push_str(&format!(..)) chains, and
format!(..).trim_end().to_string() copies are gone — one line buffer
per event, everything else writes in place, and every remaining pass
over an input is single-pass.

Each event type lives in its own module under event_render/ (assert,
guidance, sdk, fault, container, composer, log), with its tests; ansi
owns the escape handling (the ANSI tests move out of runs.rs). VTime
now renders its own display text: full_text keeps the server's exact
string, truncated_text is the truncate-never-round form.
