+++ src/runs.rs
@@ -3004,78 +3004,6 @@ mod tests {
     }
 
     #[test]
-    fn ansi_sgr() {
-        assert_eq!(strip_ansi("\x1b[1mbold\x1b[0m"), "bold");
-        assert_eq!(strip_ansi("\x1b[38;5;196mred\x1b[0m"), "red");
-        assert_eq!(strip_ansi("\x1b[38;2;255;0;0mred\x1b[0m"), "red");
-        assert_eq!(strip_ansi("\x1b[1;31;42mtext\x1b[0m"), "text");
-        assert_eq!(
-            strip_ansi(
-                "\x1b[2m2026-04-03T08:19:54Z\x1b[0m \x1b[32m INFO\x1b[0m \x1b[2mfoobar\x1b[0m\x1b[2m:\x1b[0m ready"
-            ),
-            "2026-04-03T08:19:54Z  INFO foobar: ready"
-        );
-    }
-
-    #[test]
-    fn ansi_csi_non_sgr() {
-        assert_eq!(strip_ansi("left\x1b[2Aright"), "leftright");
-        assert_eq!(strip_ansi("text\x1b[2K"), "text");
-        assert_eq!(strip_ansi("\x1b[?25hvisible"), "visible");
-        assert_eq!(strip_ansi("\x1b[?25l hidden"), " hidden");
-    }
-
-    #[test]
-    fn ansi_osc() {
-        assert_eq!(
-            strip_ansi("\x1b]0;my window title\x07text after"),
-            "text after"
-        );
-        assert_eq!(strip_ansi("\x1b]0;my title\x1b\\text after"), "text after");
-    }
-
-    #[test]
-    fn ansi_two_byte() {
-        assert_eq!(strip_ansi("\x1bcafter reset"), "after reset");
-        assert_eq!(strip_ansi("before\x1b7after"), "beforeafter");
-    }
-
-    #[test]
-    fn ansi_passthrough() {
-        let cases = [
-            "no escapes here",
-            r#"{"key": "value", "nested": {"a": [1,2,3]}}"#,
-            r#"{"url": "http://example.com/path?q=1&r=2", "count": 42}"#,
-            r#"Options { address: Some(0.0.0.0:3307), deployment: "mydb", mode: Standalone }"#,
-            r#"Settings { inner: Inner { values: [1, 2, 3] }, name: "test" }"#,
-            "[2026-04-03] [INFO] [main] started",
-            r#"path: "/nix/store/abc-pkg/bin/cmd""#,
-            r#"{"msg": "he said \"hello\""}"#,
-        ];
-        for c in cases {
-            assert_eq!(strip_ansi(c), c, "passthrough failed: {c:?}");
-        }
-    }
-
-    #[test]
-    fn ansi_mixed() {
-        assert_eq!(
-            strip_ansi("\x1b[2m{\"key\": \"value\"}\x1b[0m"),
-            r#"{"key": "value"}"#
-        );
-        assert_eq!(
-            strip_ansi("\x1b[3mOptions { mode: Standalone }\x1b[0m"),
-            "Options { mode: Standalone }"
-        );
-        assert_eq!(
-            strip_ansi(
-                "\x1b[2m2026-04-03T00:00:00Z\x1b[0m \x1b[32m INFO\x1b[0m request completed {\"status\": 200, \"latency_ms\": 42}"
-            ),
-            r#"2026-04-03T00:00:00Z  INFO request completed {"status": 200, "latency_ms": 42}"#
-        );
-    }
-
-    #[test]
     fn tracks_which_faults_are_active_based_on_vtime_and_max_duration() {
         let mut transformer = FaultAnnotator::default();
